Validators structure cleanup
