+++ src/CaptainHook.Application/ApplicationModule.cs
@@ -1,7 +1,7 @@
 ﻿using Autofac;
 using CaptainHook.Application.Handlers.Subscribers;
 using CaptainHook.Application.Infrastructure;
-using CaptainHook.Application.RequestValidators;
+using CaptainHook.Application.Validators;
 
 namespace CaptainHook.Application
 {
@@ -10,7 +10,7 @@ namespace CaptainHook.Application
         protected override void Load(ContainerBuilder builder)
         {
             var handlersAssembly = typeof(AddSubscriberRequestHandler).Assembly;
-            var validatorsAssembly = typeof(AddSubscriberRequestValidator).Assembly;
+            var validatorsAssembly = typeof(UpsertWebhookRequestValidator).Assembly;
 
             builder.RegisterMediatorInfrastructure(handlersAssembly, validatorsAssembly);
         }
